# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Quordon API
  version: 1.4.0
  description: >-
    Policy-controlled gateway for agent access to infrastructure. The current
    MVP accepts a structured QuerySpec, never raw SQL, and implements
    policy-filtered schema, table-statistics and query-shape discovery, bounded
    SELECT, bounded curated
    aggregates, and JSON EXPLAIN for MySQL 8.x
    (>= 8.0 and < 9.0). MariaDB is not supported.
    One process implements one API major version; a border proxy may add a
    public version prefix.
security:
  - basicAuth: []
servers:
  - url: /
    description: Current Quordon process; no version prefix is used internally.
...
paths:
...
  /capabilities:
    get:
      operationId: getCapabilities
      summary: Get effective capabilities
      description: Returns only profiles assigned to the authenticated principal.
      tags: [Discovery]
      responses:
        "200":
          description: Capabilities visible to the authenticated principal.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capabilities"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /schemas/{schema}/objects:
    get:
      operationId: listSchemaObjects
      summary: List allowed physical tables
      description: >-
        Returns physical BASE TABLE objects that are visible through the
        selected profile. Views and policy-denied objects are omitted before
        applying the effective result-byte limit, so hidden metadata does not
        consume the profile budget. Metadata responses are never partial. The
        query parameter set is closed: profile must occur exactly once and any
        additional or repeated query parameter and any profile that is not
        valid UTF-8 produces 400 INVALID_REQUEST. This strictness is an
        application semantic represented by the
        x-quordon-strict-query-parameters extension because OpenAPI parameter
        objects do not define a closed query-string object.
      x-quordon-strict-query-parameters: true
      tags: [Schemas]
      parameters:
        - $ref: "#/components/parameters/SchemaName"
        - $ref: "#/components/parameters/Profile"
      responses:
        "200":
          description: Complete policy-filtered object list.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectListResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/TooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
        "501":
          $ref: "#/components/responses/NotImplemented"
        "502":
          $ref: "#/components/responses/UpstreamError"
        "503":
          $ref: "#/components/responses/Unavailable"
        "504":
          $ref: "#/components/responses/Timeout"
  /schemas/{schema}/objects/{object}:
    get:
      operationId: describeSchemaObject
      summary: Describe an allowed physical table
      description: >-
        Returns only columns visible through the selected profile. A missing,
        denied, or non-table object produces the same NOT_FOUND response. The
        effective result-byte limit applies after policy filtering; metadata
        responses are never partial. The query parameter set is closed: profile
        must occur exactly once and any additional or repeated query parameter,
        or a profile that is not valid UTF-8, produces 400 INVALID_REQUEST. This
        strictness is an application semantic
        represented by the x-quordon-strict-query-parameters extension because
        OpenAPI parameter objects do not define a closed query-string object.
      x-quordon-strict-query-parameters: true
      tags: [Schemas]
      parameters:
        - $ref: "#/components/parameters/SchemaName"
        - $ref: "#/components/parameters/ObjectName"
        - $ref: "#/components/parameters/Profile"
      responses:
        "200":
          description: Policy-filtered physical table description.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectDescriptionResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/TooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
        "501":
          $ref: "#/components/responses/NotImplemented"
        "502":
          $ref: "#/components/responses/UpstreamError"
        "503":
          $ref: "#/components/responses/Unavailable"
        "504":
          $ref: "#/components/responses/Timeout"
...
  /queries/explain:
    post:
      operationId: explainSelect
      summary: Explain a structured select query
      description: >-
        Compiles an authorized QuerySpec and runs MySQL 8 EXPLAIN FORMAT=JSON.
        The source must resolve to a physical MySQL BASE TABLE; views and
        unknown objects are rejected with UNSUPPORTED_QUERY.
      tags: [Queries]
      requestBody:
        required: true
        description: A profile name and portable, structured QuerySpec.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: MySQL 8 JSON execution plan for an authorized QuerySpec.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplainResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/TooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "500":
          $ref: "#/components/responses/InternalError"
        "501":
          $ref: "#/components/responses/NotImplemented"
        "502":
          $ref: "#/components/responses/UpstreamError"
        "503":
          $ref: "#/components/responses/Unavailable"
        "504":
          $ref: "#/components/responses/Timeout"
  /queries/select:
    post:
      operationId: executeSelect
      summary: Execute a bounded structured select query
      description: >-
        Executes either the legacy bounded field-only SELECT or a separately
        authorized policy-curated keyset SELECT. The request and success
        response branches are disjoint; keyset behavior is defined by the
        executable keyset-pagination contract module.
      x-quordon-capabilities: [select, select_keyset]
      x-quordon-contract-module: ./keyset-pagination.yaml
      tags: [Queries]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./keyset-pagination.yaml#/components/schemas/SelectRequestVNext"
      responses:
        "200":
          description: Legacy or keyset SELECT result matching the request branch.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "./keyset-pagination.yaml#/components/schemas/SelectResultVNext"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "405":
          $ref: "./keyset-pagination.yaml#/components/responses/SelectMethodNotAllowed"
        "413":
          $ref: "#/components/responses/TooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "500":
          $ref: "#/components/responses/InternalError"
        "501":
          $ref: "#/components/responses/NotImplemented"
        "502":
          $ref: "#/components/responses/UpstreamError"
        "503":
          $ref: "#/components/responses/Unavailable"
        "504":
          $ref: "#/components/responses/Timeout"
  /queries/aggregate:
    post:
      operationId: executeAggregate
      summary: Execute a bounded curated aggregate query
      description: >-
        Executes one policy-authorized scalar or grouped aggregate shape over
        an InnoDB physical BASE TABLE. The complete execution invariants and
        schemas are defined in the aggregate contract module. The client never
        selects an index, SQL expression, source-work bound, or execution plan.
      x-quordon-capability: aggregate
      x-quordon-contract-module: ./aggregate.yaml
      tags: [Queries]
      requestBody:
        required: true
        description: A profile name and a strictly typed aggregate query.
        content:
          application/json:
            schema:
              $ref: "./aggregate.yaml#/components/schemas/AggregateRequest"
      responses:
        "200":
          description: Bounded aggregate rows with portable column metadata.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "./aggregate.yaml#/components/schemas/AggregateResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          $ref: "#/components/responses/TooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "500":
          $ref: "#/components/responses/InternalError"
        "501":
          $ref: "#/components/responses/NotImplemented"
        "502":
          $ref: "#/components/responses/UpstreamError"
        "503":
          $ref: "#/components/responses/Unavailable"
        "504":
          $ref: "#/components/responses/Timeout"
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Auth on every application request. Quordon does not terminate TLS; remote deployments must terminate HTTPS at an external ingress.
  headers:
    RequestId:
      description: Server-generated request correlation identifier.
      schema:
        type: string
  parameters:
    Profile:
      name: profile
      in: query
      required: true
      description: >-
        Assigned server-side policy profile. On metadata operations this
        parameter must occur exactly once; repeated occurrences are rejected
        with 400 INVALID_REQUEST, as is a percent-decoded value that is not
        valid UTF-8.
      schema:
        type: string
        minLength: 1
    SchemaName:
      description: Portable logical schema identifier.
      name: schema
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
    ObjectName:
      description: Portable physical table identifier.
      name: object
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Identifier"
  responses:
    BadRequest:
      description: Malformed JSON or invalid HTTP request.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid HTTP Basic credentials.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        WWW-Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Policy denied the profile, operation, resource, or field.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The object is missing, denied by policy, or not a physical table.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooLarge:
      description: A configured request or result size limit was exceeded.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    UpstreamError:
      description: MySQL rejected or failed the generated statement.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: An internal service error occurred.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotImplemented:
      description: The selected adapter does not implement this endpoint capability.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unavailable:
      description: Concurrency capacity is exhausted or a required dependency is unavailable.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Capabilities:
      description: Effective MVP capabilities for the authenticated principal.
      type: object
      additionalProperties: false
      required: [api_version, service_version, policy_version, profiles]
      properties:
        api_version:
          type: string
          const: "1"
        service_version:
          type: string
        policy_version:
          type: string
        profiles:
          type: array
          items:
            $ref: "#/components/schemas/ProfileCapability"
...
    ObjectListResult:
      description: Complete list of physical tables visible through one profile.
      type: object
      additionalProperties: false
      required: [policy_profile, policy_version, datasource, adapter, schema, objects]
      properties:
        policy_profile:
          type: string
        policy_version:
          type: string
        datasource:
          type: string
        adapter:
          type: string
          const: mysql8
        schema:
          $ref: "#/components/schemas/Identifier"
        objects:
          type: array
          items:
            $ref: "#/components/schemas/SchemaObject"
...